requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
django-heroku==0.3.1
gunicorn==20.1.0
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.2
whitenoise==5.3.0
Django==4.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
hazm==0.7.0
PyJWT==2.3.0
pyparsing==3.0.6
pyrsistent==0.18.0
QtPy==2.0.0
radiojavanapi==0.2.2
requests==2.25.1
annoy
numpy
tensorflow==2.5.1
scipy==1.7.2
resampy==0.2.1
six==1.15.0
maxfw
tf_slim
pydub